+++ database/migrations/2025_03_12_043128_create_general_care_plans_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('general_care_plans', function (Blueprint $table) {
             $table->increments('general_care_plan_id');
-            $table->unsignedBigInteger('beneficiary_id')->unique()->after('general_care_plan_id');
-            $table->unsignedBigInteger('care_worker_id')->after('beneficiary_id');
+            $table->integer('beneficiary_id')->unique()->after('general_care_plan_id');
+            $table->integer('care_worker_id')->after('beneficiary_id');
             $table->text('emergency_plan')->after('care_worker_id');
             $table->timestamps();
